requirements.txt
django
django-proxy
mysqlclient
